+++ packages/swapkit-core/src/client/index.ts
@@ -34,6 +34,7 @@ import {
   FeeOption,
   TCAvalancheDepositABI,
   TCEthereumVaultAbi,
+  WalletOption,
 } from '@thorswap-lib/types';
 
 import {
@@ -182,23 +183,20 @@ export class SwapKitCore {
 
   getAddress = (chain: Chain) => this.connectedChains[chain]?.address || '';
 
+  getBalance = async (chain: Chain, refresh?: boolean) => {
+    if (!this.connectedChains[chain]?.address) return [];
+    if (!refresh) return this.connectedChains[chain]?.balance || [];
+    const chainData = await this.getWalletByChain(chain);
+
+    return chainData?.balance || [];
+  };
+
   getExplorerAddressUrl = (chain: Chain, address: string) =>
     getExplorerAddressUrl({ chain, address });
 
   getExplorerTxUrl = (chain: Chain, txHash: string) => getExplorerTxUrl({ chain, txHash });
 
   getWallet = <T extends Chain>(chain: Chain) => this.connectedWallets[chain] as WalletMethods[T];
-
-  isAssetApproved = (asset: AssetEntity) => this._approve<boolean>({ asset }, 'checkOnly');
-
-  isAssetApprovedForContract = (
-    asset: AssetEntity,
-    contractAddress: string,
-    amount?: AmountWithBaseDenom,
-  ) => this._approve<boolean>({ asset, amount, contractAddress }, 'checkOnly');
-
-  validateAddress = ({ address, chain }: { address: string; chain: Chain }) =>
-    this.getWallet(chain)?.validateAddress?.(address);
 
   getWalletByChain = async (chain: Chain) => {
     const address = this.getAddress(chain);
@@ -216,8 +214,25 @@ export class SwapKitCore {
         ),
     );
 
+    this.connectedChains[chain] = {
+      address,
+      balance,
+      walletType: this.connectedChains[chain]?.walletType as WalletOption,
+    };
+
     return { ...(this.connectedChains[chain] || {}), balance };
   };
+
+  isAssetApproved = (asset: AssetEntity) => this._approve<boolean>({ asset }, 'checkOnly');
+
+  isAssetApprovedForContract = (
+    asset: AssetEntity,
+    contractAddress: string,
+    amount?: AmountWithBaseDenom,
+  ) => this._approve<boolean>({ asset, amount, contractAddress }, 'checkOnly');
+
+  validateAddress = ({ address, chain }: { address: string; chain: Chain }) =>
+    this.getWallet(chain)?.validateAddress?.(address);
 
   transfer = async (params: CoreTxParams & { router?: string }) => {
     const chain = params.assetAmount.asset.L1Chain;